+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 current dev branch
 
+## [0.5.2] - 2020 - 02 -15
+
+### Added
+
+- location utils
+    - geolocate
+    - reverse geolocate
+    - get timezone
+    
 ## [0.5.1] - 2020 - 02 -04
 
 ### Fixed